refactor(ui): split ui.py into a ui/ package with a re-exporting facade
